app/ai/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
scikit-learn
openai
